requirements.txt
langchain
chromadb
huggingface-hub
openai
pyPDF
python-dotenv

